1.加入总电控制，在上面的继电器二，用的是继电器控制继电器。 2.解决了计数闪动问题，是状态检测那里，加上计数是32位的数据，而原来程序只给了16位，现在修改好了。 3.加入错误编号显示，在有错误的时候显示其编号。

+++ HARDWARE/MODBUS/RS485.h
@@ -7,10 +7,9 @@
 void Timer7_Init(void);
 void RS485_Init(void);
 void RS485_Service(void);
-extern u32 Cylinder_Data[20];
+extern u16 Cylinder_Data[20];
 extern u32 Connect_Data[20];
 extern u32 Status_Data[20];
-
 extern vu32 *Modbus_InputIO[100];
 #endif
